sql/requirements.txt
pytest==8.3.3
flake8==7.1.1
discord.py==2.3.2
python-dotenv==1.0.1
supabase>=2.15.0
requests>=2.31.0
beautifulsoup4>=4.12.3
pandas>=2.2.3setuptools
supabase>=2.0.0
setuptools
setuptools
psycopg2-binary
